requirements.txt
# GRPC / Protobuff related
grpcio>=1.16.1
grpcio-tools>=1.16.1
